hospital-management-js/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prisma==5.17.0
python-multipart==0.0.6
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiosmtplib==3.0.1
email-validator==2.1.0
slowapi==0.1.9
redis==5.0.1
